requirements.txt
pycryptodomex
requests